[feat] 관리자용 워크스페이스 조회 API에 use resource 추가

+++ module-k8s/src/main/java/com/xiilab/modulek8s/facade/dto/WorkspaceTotalDTO.java
@@ -21,9 +21,9 @@ public class WorkspaceTotalDTO {
 	private String name;
 	private String resourceName;
 	private String description;
-	private int limitCPU;
-	private int limitMEM;
-	private int limitGPU;
+	private float limitCPU;
+	private float limitMEM;
+	private float limitGPU;
 	private AgeDTO age;
 	private int runningCnt;
 	private int endCnt;
